+++ libraries/tools/kotlin-gradle-plugin/src/main/kotlin/org/jetbrains/kotlin/gradle/plugin/KotlinPlugin.kt
@@ -106,6 +106,7 @@ abstract class KotlinSourceSetProcessor<T : AbstractCompile>(
         javaBasePlugin.configureForSourceSet(sourceSet, kotlinTask)
         kotlinTask.description = taskDescription
         kotlinTask.source(kotlinDirSet)
+        mapKotlinTaskProperties(project, kotlinTask)
     }
 }
 
@@ -376,6 +377,7 @@ open class KotlinAndroidPlugin @Inject constructor(val scriptHandler: ScriptHand
 
             val kotlinTaskName = "compile${variantDataName.capitalize()}Kotlin"
             val kotlinTask = tasksProvider.createKotlinJVMTask(project, kotlinTaskName)
+            mapKotlinTaskProperties(project, kotlinTask)
 
             kotlinTask.extensions.extraProperties.set("defaultModuleName", "${project.name}-$kotlinTaskName")
             if (kotlinOptions != null) {