+++ libraries/tools/kotlin-gradle-plugin/src/test/kotlin/org/jetbrains/kotlin/gradle/BaseGradleIT.kt
@@ -70,6 +70,7 @@ abstract class BaseGradleIT {
     data class BuildOptions(
             val withDaemon: Boolean = false,
             val daemonOptionSupported: Boolean = true,
+            val incremental: Boolean? = null,
             /**
              * @see [ThreadTracker]
              */
@@ -201,8 +202,8 @@ abstract class BaseGradleIT {
     private fun Project.createBuildCommand(params: Array<out String>, options: BuildOptions): List<String> =
             createGradleCommand(createGradleTailParameters(options, params))
 
-    protected fun Project.createGradleTailParameters(options: BuildOptions, tasks: Array<out String> = arrayOf()): List<String> =
-            tasks.toMutableList().apply {
+    private fun Project.createGradleTailParameters(options: BuildOptions, params: Array<out String> = arrayOf()): List<String> =
+            params.toMutableList().apply {
                 add("--stacktrace")
                 add("--${minLogLevel.name.toLowerCase()}")
                 if (options.daemonOptionSupported) {
@@ -213,6 +214,8 @@ abstract class BaseGradleIT {
                 if (options.assertThreadLeaks) {
                     add("-P${ThreadTracker.ASSERT_THREAD_LEAKS_PROPERTY}=true")
                 }
+                options.incremental?.let { add("-Pkotlin.incremental=$it") }
+            }
 
             }